Bug fixed by Daniel Minder
Before a node is removed from the synchronization group it should be
assured that all other nodes are "later" than this node. Otherwise,
the goal variable is decremented but the numpassed variables of the
wait list are not adjusted which may lead to early return from
waitForLink(). Actually, this happens when Low Power Listening puts
a node to sleep mode just before the end of simulation time.
Finally, this results in an array of out bound exception in the
medium implementation.