Bugs fixed by Daniel Minder
For Timer1, the interupt bits are distributed on four
registers (TIMSK, ETIMSK, TIFR, ETIFR) the ATmega128 processor. This is
now correctly implemented. For this, changes were also necessary in
ATMegaFamily. Also, the InputCapturePin has a flag that determines
if an interrupt should be triggered on a rising or falling edge. I
thought this missing feature caused an error but this was not
true. But anyway, it's implemented now, so you get it :) A change in
ATMega128 was necessary for this.